chore: rename after beta merge
